+++ clogging/src/features/Comment/types.ts
@@ -11,21 +11,57 @@ export interface Comment {
   replies?: Comment[];
 }
 
-export interface CreateCommentDto {
-  postId: string;
-  nickname: string;
-  password: string;
-  content: string;
-  isPrivate: boolean;
-  isAuthor: boolean;
-  parentCommentId?: string;
+export interface CommentWithReplies extends Omit<Comment, 'replies'> {
+  replies?: CommentWithReplies[];
 }
 
-export interface UpdateCommentDto {
-  id: string;
+export interface commentItemProps {
+  comment: CommentWithReplies;
+  level?: number;
+  onEdit: (commentId: string) => void;
+  onDelete: (commentId: string) => Promise<void>;
+  onReply: (commentId: string) => void;
+  editingCommentId: string | null;
+  editingContent: string;
+  editingIsPrivate: boolean;
+  onEditContentChange: (content: string) => void;
+  onEditPrivateChange: (isPrivate: boolean) => void;
+  onEditSubmit: (commentId: string) => Promise<void>;
+  onEditCancel: () => void;
+  replyingToId: string | null;
+  onReplySuccess: () => void;
+  onReplyCancel: () => void;
   postId: string;
-  content: string;
+}
+
+export interface commentFormProps {
+  postId: string;
+  onSuccess: () => void;
+  commentId?: string;
+  initialData?: {
+    nickname: string;
+    content: string;
+    isPrivate: boolean;
+  };
+  mode?: 'create' | 'edit';
+  parentCommentId?: string;
+  defaultNickname?: string;
+  hideFields?: boolean;
+  defaultIsPrivate?: boolean;
+}
+
+export interface InputCommentProps {
+  value: string;
+  onChange: (value: string) => void;
+  initialData?: Comment;
+  defaultNickname?: string;
+  postId: string;
+  commentId: string;
+  onSuccess?: () => void;
+  mode: 'create' | 'edit';
+}
+
+export interface PrivateCommentProps {
   isPrivate: boolean;
-  nickname: string;
-  password: string;
+  onChange: (value: boolean) => void;
 }
